cobertura con sonar, 0 bugs,1 code smell, falta pruebas services

+++ src/test/java/org/bedu/atko/repository/IClientRepositoryTest.java
@@ -56,7 +56,7 @@ class IClientRepositoryTest {
     @DisplayName("Guarda un cliente")
     void save(){
         client = clientRepository.save(client);
-        assertNotNull(client.getId());
+        assertNotNull((Long) client.getId());
     }
 
     @Test
